Converted Map Structure to MVVM
