+++ Dating App/App.xaml
@@ -1,8 +1,9 @@
 ﻿<?xml version = "1.0" encoding = "UTF-8" ?>
 <Application xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="Dating_App.App"
              xmlns:local="clr-namespace:Dating_App"
-             x:Class="Dating_App.App">
+             xmlns:authentication="clr-namespace:Dating_App.MVVM.Views.Authentication">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
