+++ Dating App/MVVM/Views/Messages.xaml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Dating_App.Messages"
+             x:Class="Dating_App.MVVM.Views.Messages"
              Title="Messages"
              Shell.TabBarIsVisible="True">
