+++ Dating App/MVVM/Views/Navigationbar.xaml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <TabbedPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Dating_App.Pages.Navigationbar"
-             xmlns:local="clr-namespace:Dating_App"
-             xmlns:authentication="clr-namespace:Dating_App.Pages.Authentication"
+             x:Class="Dating_App.MVVM.Views.Navigationbar"
+             xmlns:local="clr-namespace:Dating_App.MVVM.Views"
+             xmlns:authentication="clr-namespace:Dating_App.MVVM.Views.Authentication"
              xmlns:android="clr-namespace:Microsoft.Maui.Controls.PlatformConfiguration.AndroidSpecific;assembly=Microsoft.Maui.Controls"
              android:TabbedPage.ToolbarPlacement="Bottom">
