+++ Dating App/MVVM/Views/Navigationbar.xaml.cs
@@ -1,11 +1,10 @@
-using Dating_App.Models.Data;
-using Dating_App.Models;
+using Dating_App.MVVM.Models.Data;
+using Dating_App.MVVM.Models;
 
-namespace Dating_App.Pages;
+namespace Dating_App.MVVM.Views;
 
 public partial class Navigationbar : TabbedPage
 {
-	
     public Navigationbar()
 	{
 		InitializeComponent();
