+++ Dating App/MVVM/Views/Quiz.xaml.cs
@@ -1,0 +1,13 @@
+
+using Dating_App.MVVM.Models;
+using Dating_App.MVVM.Models.Data;
+
+namespace Dating_App.MVVM.Views;
+
+public partial class Quiz : ContentPage
+{
+    public Quiz()
+	{
+		InitializeComponent();
+	}
+}
