+++ Dating App/MVVM/Views/Swipe.xaml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Dating_App.Swipe"
+             x:Class="Dating_App.MVVM.Views.Swipe"
              Title="Swipe"
              Shell.TabBarIsVisible="True">
     <Grid RowDefinitions="Auto, *">
