--- Dating App/Pages/Quiz.xaml.cs
@@ -1,13 +1,0 @@
-
-using Dating_App.Models;
-using Dating_App.Models.Data;
-
-namespace Dating_App;
-
-public partial class Quiz : ContentPage
-{
-    public Quiz()
-	{
-		InitializeComponent();
-	}
-}
